[S07P12D105-204] [BE] feat : 로그인 완료 시 반환 정보 추가

See merge request s07-webmobile1-sub2/S07P12D105!53

+++ BE/sellerB/src/main/java/backend/sellerB/entity/Consultant.java
@@ -22,7 +22,7 @@ public class Consultant implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "consultant_seq")
-    private int consultantSeq;
+    private Long consultantSeq;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_group_seq")
     private ProductGroup productGroup;
